Rebase moonray-4848-v3 (#1805)
* Rebase moonray-4848-v3

* Add atomicStore function to Atomic.h

An atomicStore function was added to the Atomic.h file. This function allows to store a value in atomic way with a given 'std::memory_order' and thus providing more control over how a program performs in multi-threaded environment. The function is templatized, which increases its versatility and applicability, and it uses '__atomic_store_n' primitive.

* Add Atomic.h to public headers and update atomicStore function

Atomic.h has been added to the public_headers list in SConscript, as it is necessary for the general API. Meanwhile, in Atomic.h, the atomicStore function was updated. The '__atomic_store_n' function was replaced with the '__atomic_store' function to accommodate generic types.

* Remove lock-free check in atomicStore